pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "stock-telnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "websockets>=15.0.1",
    "yfinance>=0.2.54",
]

[tool.setuptools]
package-dir = { "" = "telnet_server" }
